Move r22593, r22594, r22604, r22607, r22612, r22649, r22662, r22670 and r22686.
Check whether file is opened on network file-system. If file does not exist, check the directory it lives in... Maybe used by caller, trying to open mmap() on NFS, Lustre or Panasas (thanks Sam). For now, this is used to warn about the usage of mmap on such FS.

This commit closes trac:2256.

This commit was SVN r22694.

The following SVN revision numbers were found above:
  r22593 --> open-mpi/ompi@c161cf5fa45c7ee4ed54043278993ea24f15a7f5
  r22594 --> open-mpi/ompi@583bb42739304e05df7c53e93f8d1142740b871d
  r22604 --> open-mpi/ompi@ea4de16561d0f74d1f7f2dd7cabb2c3bff8a66d6
  r22607 --> open-mpi/ompi@ecbd530a77879e2639fdeec6a2dca91401b50c54
  r22612 --> open-mpi/ompi@0538b1a9481d2cab384d31633931c1b1190f99fe
  r22649 --> open-mpi/ompi@a46cecf4f21483cf3e405f8614d8bb3618d03f72
  r22662 --> open-mpi/ompi@8dd87def778790678abee29fa1039053023faaae
  r22670 --> open-mpi/ompi@548d6f7c613276fb5413a433f64c705fc1b656ac
  r22686 --> open-mpi/ompi@cfe37fb5a1cb1cfef8b551f0ee086b8b8335d590

The following Trac tickets were found above:
  Ticket 2256 --> https://svn.open-mpi.org/trac/ompi/ticket/2256