Per RFC initially started here:
    http://www.open-mpi.org/community/lists/devel/2010/02/7496.php

Increase the required versions of AM, AC, and LT:

 * Autoconf: 2.65
 * Automake: 1.11.1
 * Libtool: 2.2.6b

And therefore removed a bunch of patches that we used to apply to make
older versions of these tools work.

Also updated the HACKING document to match these version numbers,
specifically mentioned Mercurial in a few places, and removed some
outdated language about running autogen.sh in subdirectories.

This commit was SVN r22896.