From r22675: Fixes the problem where the rcache and core memory allocation can deadlock itself.
From r22863:
There is a lengthy discussion about this patch on #2322.
The short version is that this patch consolidates two pieces of code
that call the back-end munmap and ensures that (if dlsym is used) the
corresponding dlsym is only invoked once and that the variable holding
the result is volatile.

From r22867:
Update copyrights.

This commit closes trac:2358.

This commit was SVN r22905.

The following SVN revision numbers were found above:
  r22675 --> open-mpi/ompi@bccafbb5dfb5772e9d87fc03a41aaf156b026491
  r22863 --> open-mpi/ompi@136f926fd1054139b361520304a2f498abe339b9
  r22867 --> open-mpi/ompi@59126b1e0b4dfe9ba2cded063b64ff3d033cb17d

The following Trac tickets were found above:
  Ticket 2358 --> https://svn.open-mpi.org/trac/ompi/ticket/2358