Fixes trac:2369: Please move r22927 to 1.5 to fix a processor affinity issue
r22927
================================================================================
modify the OPAL_PAFFINITY_PROCESS_IS_BOUND macro to search the cpuset for
the maximum possible number of cpus rather than just the number of cpus
currently online.  This corrects a problem where mpi_paffinity_alone was
not working properly on systems in which there can be cpu namespaces with
holes, such as on ppc64 with smt off (as discussed in #2365).

This commit was SVN r23003.

The following SVN revision numbers were found above:
  r22927 --> open-mpi/ompi@58a9aeff5ac845f84f6a25a6e681e0ba5e7d9b9e

The following Trac tickets were found above:
  Ticket 2369 --> https://svn.open-mpi.org/trac/ompi/ticket/2369