Fixes trac:2371: Move r23014 to v1.5 (WAS: mpirun segv's on ppc64 when built optimized with gcc-4-3-2 (SLES11))
The "mr %3 %0" is not needed and in fact corrupts v. Note that t already
holds the updated value. The mr inclusion was an accidentaly hangover
from the updated assembler version of this function. Note that the
assembler version is correct and does need the mr.

This commit was SVN r23026.

The following SVN revision numbers were found above:
  r23014 --> open-mpi/ompi@cab7982c7e45a6ff93cc8ae302959e76825cbba6

The following Trac tickets were found above:
  Ticket 2371 --> https://svn.open-mpi.org/trac/ompi/ticket/2371