Fixes trac:2337: Move libltdl preopen patch to v1.5
r22806
Add a patch to libltdl so that the preopen loader doesn't mask a real
error if the file-based dloader (e.g., dlopen) fails to load a DSO for
a complex reason (e.g., unresolved symbol).

This has been widely reported upstream to the libltdl maintainers - a
general solution is difficult.  This is very definitely an
OMPI-specific solution.  Since our embedded libltdl is hidden behind
visibility flags, that's ok.

Note that this is a change to autogen.sh, but this commit does not
force re-running autogen.sh.  You'll just get the new functionality
the next time you re-run autogen.sh.

r22816
Oops -- use a name that matches the other patches in this dir and
matches the name in Makefile.am.

r22823
Fix for the libltdl preopen commit. So now autogen will progress without emiting an error.

{{{
grep: configure: No such file or directory
./autogen.sh: line 495: config/libltdl-preopen-error.patch: No such file or directory
}}}

This commit was SVN r23223.

The following SVN revision numbers were found above:
  r22806 --> open-mpi/ompi@9f1c699f360b91493523aba696893fe7af775773
  r22816 --> open-mpi/ompi@d0d1cd90dbd9d92350079d1ad1ad85fa08c0988e
  r22823 --> open-mpi/ompi@f3848f169b504b1819609f38b6637b13ad5b61f4

The following Trac tickets were found above:
  Ticket 2337 --> https://svn.open-mpi.org/trac/ompi/ticket/2337