Fixes trac:2442: Move dlopen test fix to v1.5
r23270
Add in 2 cases for where this test may be skipped:

 1. If opal wasn't built with libltdl support
 1. If opal was built statically (i.e., dlopen='' in the .la file)

r23280
Search for "dlname", not "dlopen".  This value will  be filled in if
there is a DSO to open.

r23281
Use the right number of characters for the strncmp.  Thanks to Brad
for catching that!

This commit was SVN r23294.

The following SVN revision numbers were found above:
  r23270 --> open-mpi/ompi@b620e63bdcbbf5cea64d3baea8ff81ba98ff4733
  r23280 --> open-mpi/ompi@cdc5541cb02d74621b1dd30ea075d6e7a0056d91
  r23281 --> open-mpi/ompi@ea05c73cfc059d1621761ba0bda7f8ace56078cf

The following Trac tickets were found above:
  Ticket 2442 --> https://svn.open-mpi.org/trac/ompi/ticket/2442