Fixes trac:2450.
Remove xgrid support.  It's broken, we known it's broken, there's
currently no one available to fix it (and no one has been available
for at least a year).  So let's not pretend we have xgrid support; we
don't.  :-(

If anyone would like to fix this situation, please contact us ASAP -
we would love to re-enable xgrid support (all the code is available in
SVN history; it's easy to resurrect).

This commit was SVN r23311.

The following Trac tickets were found above:
  Ticket 2450 --> https://svn.open-mpi.org/trac/ompi/ticket/2450