Fixes trac:2470: Move coll tuned dynamic rules fix to v1.5
r23361
Ensure that the com_rules[] array entries are initialized to NULL in
case individual entries aren't used, but dynamic rules are enabled
(i.e., at least one or more of them are not NULL, meaning that they'll
all be assumed to be either NULL or a valid value).

This commit was SVN r23395.

The following SVN revision numbers were found above:
  r23361 --> open-mpi/ompi@87e17a41daeb45688d572f618efe2c0b7cf78609

The following Trac tickets were found above:
  Ticket 2470 --> https://svn.open-mpi.org/trac/ompi/ticket/2470