Fixes trac:2493: Remove undeclared variable from ptmalloc2 munmap
Submitted by Manuel Prinz, reviewed by Jeff.

This only happens on systems that have dlsym but do not have __munmap or syscall (e.g., at least one of the BSDs), which is probably why we didn't notice. :-(

This is not an issue for the trunk or v1.5 branches.

NOTE: hand-patched

This commit was SVN r23507.

The following Trac tickets were found above:
  Ticket 2493 --> https://svn.open-mpi.org/trac/ompi/ticket/2493