Fixes trac:2534: Move MPI_OP string name changes to v1.5
r23600
r20280 introduced the op framework and changed all the back-end op
string names from MPI_<foo> to MPI_OP_<foo>.  While these names are
OMPI-internal-only (i.e., not exposed to MPI applications), this
change is a difference between the released 1.3/1.4 series.

The Voltaire FCA library uses these strings for its own internal
purposes; since the names changed between the 1.3/1.4 series and the
upcoming 1.5 series, it caused a problem for the FCA library.  They
volunteered to put in a hot fix in FCA, but it seems to me that we
shouldn't change the names to begin with -- there was no real reason
to change them to MPI_OP_<foo>.  So this commit changes them back to
MPI_<foo>. 

This commit was SVN r23621.

The following SVN revision numbers were found above:
  r20280 --> open-mpi/ompi@4d8a18745038f41c8f1bdd4e92e07b5ead877656
  r23600 --> open-mpi/ompi@e6f0422f7c2d72c3791cd6614aecf0ca472522df

The following Trac tickets were found above:
  Ticket 2534 --> https://svn.open-mpi.org/trac/ompi/ticket/2534