Short version:
This script automates the tedious task of updating copyright notices
in the tops of OMPI/ORTE/OPAL source files before committing back to
the respository.  Set the environment variable
OMPI_COPYRIGHT_SEARCH_NAME to a short (case-insensitive) name that
indicates your copyright line (e.g., "cisco"), and set the env
variable OMPI_COPYRIGHT_FORMAL_NAME with your organization's formal
name and copyright statement (e.g., "Cisco Systems, Inc.  All rights
reserved.") before running the script.

More details:

This is a simple script to traverse the tree looking for added and
changed files (via "svn st ." or "hg st .", depending on what meta
directory is found in this tree).  Note that the search starts in
the current directory -- not the top-level directory.

All added and changed files are examined.  If the special
"$COPYRIGHT$" token is found, then lines above that token are
examined to find the "search" copyright name.

- If the search name is found, that line is examined to see if the
  current year is in the copyright year range.  If it is not, the line
  is modified to include the current year.
- If the search name is not found, a new line is created in the
  copyright block of the file using the formal name and the current
  year.

NOTE: this script currently doesn't handle multi-line copyright
statements, such as:
 
Copyright (c) 2010 University of Blabbityblah and the Trustees of
                   Schblitbittyboo.  All rights reserved.

Someone could certainly extend this script to do so, if they cared
(my organizations' copyright fits on a single line, so I wasn't
motivated to handle the multi-line case :-) ).

This commit was SVN r23742.