Fixes trac:2672: Move MPI_Init_thread C/R fix to v1.4 branch
r24194
MPI_Init_thread is used in place of MPI_Init, so for the checkpoint/restart functionality it must correctly init the C/R functionality instead of simply making a critical section. This allows the C/R thread to be started properly.

Thanks to Takayuki Seki for finding this bug.


r24200
move the INIT to after the error handler, so it matches MPI_INIT. Thanks to Jeff for catching this

This commit was SVN r24225.

The following SVN revision numbers were found above:
  r24194 --> open-mpi/ompi@0b514e234be5543787b2acc710f898c0ccbf79ae
  r24200 --> open-mpi/ompi@2bdff63e6ffbfadc74f22ccb2c58a3e63b40f1ec

The following Trac tickets were found above:
  Ticket 2672 --> https://svn.open-mpi.org/trac/ompi/ticket/2672