Fixes trac:2671: Move MPI_Init_thread C/R fix to v1.5 branch
r24194
MPI_Init_thread is used in place of MPI_Init, so for the checkpoint/restart functionality it must correctly init the C/R functionality instead of simply making a critical section. This allows the C/R thread to be started properly.

Thanks to Takayuki Seki for finding this bug.

This commit was SVN r24238.

The following SVN revision numbers were found above:
  r24194 --> open-mpi/ompi@0b514e234be5543787b2acc710f898c0ccbf79ae

The following Trac tickets were found above:
  Ticket 2671 --> https://svn.open-mpi.org/trac/ompi/ticket/2671