Fixes trac:2406: Remove "linux" component, add "hwloc" paffinity component to v1.5
The first part of this commit was copying and backporting the hwloc
paffinity component from the trunk@24217.  After that, more SVN r
numbers were applied: 

r23124
Clean up some comments; make paffinity/base/base.h comments agree with
paffinity/paffinity.h. 

r23125
Make the hwloc paffinity component available for everyone.  hwloc
supports a wide variety of operating systems and platforms; see the
opal/mca/paffinity/hwloc/hwloc/README file for details.

This component includes an embedded copy of hwloc, currently based on
hwloc-1.0rc6.  But note that hwloc is properly SVN imported into the
/vendor branch, so it will be easy to update when 1.0 GA is released.
Note that the hwloc tree embedded in opal/mca/paffinity/hwloc/hwloc is
identical to a hwloc distribution tarball, except that much of the
documentation was rm -rf'ed (because we don't need it for the embedded
case).

Since the paffinity framework currently does not understand hardware
threads, the hwloc component compensates for this by identifying cores
by the "first" hardware thread on that core.  Hopefully we'll update
paffinity someday to understand hardware threads.  :-)

configure grew a --with-hwloc option, analogous to what we do for many
other external libraries that OMPI supports.  However, there's a new
feature: due to the request of several distros, OMPI can be configured
to build with its internal copy of hwloc or with an external copy of
hwloc (e.g., a system-installed hwloc).

 1. If --with-hwloc is not specified, Open MPI will try to use its
    internal copy (but silently fail/ignore hwloc if that fails).
 1. If --with-hwloc=<dir> is supplied, Open MPI looks for hwloc
    support in <dir> (and --with-hwloc-libdir=<dir>, if specified).
 1. If --with-hwloc=external is supplied, Open MPI will look for hwloc
    in a compiler/linker default external location.
 1. If --with-hwloc=internal is supplied, Open MPI will use its
    internal copy of hwloc.

Some of OMPI's main configury had to be slightly re-arranged in the
bootstrapping phase to accomodate hwloc's configry needs.

r23126
Remove the "linux" paffinity component (i.e., the one that was based
on the now-defunct PLPA) -- the new hwloc component supersedes it.  

So long, PLPA -- we loved ya!

This commit was SVN r24244.

The following SVN revision numbers were found above:
  r23124 --> open-mpi/ompi@ca6d95a9c8cf9ccefe36f9a1b0bdd7f36bb72f0e
  r23125 --> open-mpi/ompi@3129ccd9ec0e0fdeedcccb7d62bad9fc68051dcf
  r23126 --> open-mpi/ompi@21178f937918e3d0413a9867eb3ef91f07ee0799

The following Trac tickets were found above:
  Ticket 2406 --> https://svn.open-mpi.org/trac/ompi/ticket/2406