Fixes trac:2676: Fix incorrect text in MPI_FILE_SET_VIEW man page
r24201
Correct an error in MPI_FILE_SET_VIEW man page: the fh does ''not''
have to have an identical ''value'' (it must be the same file handle,
but that means nothing about its actual value).  But the datarep must
have an identical value.  Additionaly, the etype does not need to be
identical, but the extent of all the etypes supplied must be
identical.

See MPI-2.2 p401-402 for further details.

This commit was SVN r24263.

The following SVN revision numbers were found above:
  r24201 --> open-mpi/ompi@7648c3602359ffa1d95e9c3994e85b02cb6d25c9

The following Trac tickets were found above:
  Ticket 2676 --> https://svn.open-mpi.org/trac/ompi/ticket/2676