As noted on the mailing list by Gabriele Fatigati (http://www.open-mpi.org/community/lists/users/2011/01/15427.php), the --tv (and friends) switches to mpirun would effectively munge the orterun command line together and then split it apart again before exec'ing the underlying debugger.  We would therefore lose multi-token argv[x] value and split them into multiple tokens.  For example:
   mpirun --tv -np 2 a.out "foo bar"

would get launched with "foo" and "bar" as separate arguments; not one
argument.  This was due to the underlying code joining the argv into a
single string and then re-splitting it.  This commit removed the argv
join; it now does the parsing and re-jigering of the argv by only
looking at each individual argv item; multi-word tokens like "foo bar"
will never be split into separate tokens.

This commit was SVN r24322.