Fixes trac:2750: affinity mpiext is broken
r22985
Establish a method by which a process knows if it has been bound by mpirun. This helps resolve a problem where a process gets "bound" to all available resources, which looks to the opal paffinity system as "not bound". This can cause mpi_init to attempt to "bind" the process itself, causing unintended behavior.


r22998
Provide two new opal paffinity utilities for printing a hex representation of the cpu set and parsing that string back into a cpu set on the other end. Also add a new MCA param for passing the cpu set applied to a process during launch down to that process so it can know what we attempted to do.

All to be used in some new MPI extensions provided by Jeff so that users can easily query their binding situation.

This commit was SVN r24530.

The following SVN revision numbers were found above:
  r22985 --> open-mpi/ompi@4d06125a3349324ab16aa5c5e8d756ae73116b9f
  r22998 --> open-mpi/ompi@86228aee38e1e6b25ed4b384d42bd7a17b7a61f2

The following Trac tickets were found above:
  Ticket 2750 --> https://svn.open-mpi.org/trac/ompi/ticket/2750