Fixes trac:2756: Move MPI_Probe return code fix to v1.4
r24551
Fix return code from MPI_Probe and MPI_Iprobe.

Instead of returning MPI_SUCCESS every time they are called regardless of the status of the call, they should return a value representative of the action. So similar to MPI_Wait/MPI_Test they will return MPI_SUCCESS if the action was successfull, or the value that matches status.MPI_ERROR for the operation if it is unsuccessful.

This was discussed on the [http://www.open-mpi.org/community/lists/devel/2011/03/9109.php ompi-devel list]

This commit was SVN r24555.

The following SVN revision numbers were found above:
  r24551 --> open-mpi/ompi@045035963af04748f1bf16592742c5ac57965baa

The following Trac tickets were found above:
  Ticket 2756 --> https://svn.open-mpi.org/trac/ompi/ticket/2756