Fixes trac:2728: Move THREAD_MULTIPLE fixes to v1.5
r22841
Merge in the modified thread configure option branch per today's telecon.

Remove the --enable-progress-threads option as this is no longer functional, and hardcode OPAL_ENABLE_PROGRESS_THREADS to 0.

Replace the --enable-mpi-threads option with --enable-mpi-thread-multiple as this is clearer as to meaning. This option automatically turns "on" opal thread support if it wasn't already so specified. If the user specifies --disable-opal-multi-threads --enable-mpi-thread-multiple, we will error out with a message

Add a new --enable-opal-multi-threads option that turns "on" opal thread support without doing anything wrt mpi-thread-multiple

This commit was SVN r24565.

The following SVN revision numbers were found above:
  r22841 --> open-mpi/ompi@b400b84162a9737d9c5ba58b4f993326aca13a8c

The following Trac tickets were found above:
  Ticket 2728 --> https://svn.open-mpi.org/trac/ompi/ticket/2728