Fixes trac:2765: fixes comm_Create segfault with GROUP_EMPTY
r24595
avoid calling omp_dpm.mark_dyncomm if the size of the local communicator
is zero. The routine assumes that at least one process is available in the
group, which lead to a segfault when creating communicators with GROUP_EMPTY.

r24596
fix a formatting issue 

This commit was SVN r24604.

The following SVN revision numbers were found above:
  r24595 --> open-mpi/ompi@ad9f793ce461898451143a7d8adbd36d0a65a4b9
  r24596 --> open-mpi/ompi@725a0d2100d48d16b59a959562b990d54c2e0823

The following Trac tickets were found above:
  Ticket 2765 --> https://svn.open-mpi.org/trac/ompi/ticket/2765