Fixes trac:2775: CMR: fix debugger mapping when -npernode is used
There is a bug in the load_balance mapper that appears when using both -mca debugger mpirx and -npernode. In this case the mapper policy specified by mpirx is ignored and debugger daemons are not mapped by node. The problem is the load_balance module incorrectly uses the job global value of npernode (also nperboard, and npersocket) when it should instead use the value in jdata->map. The attached patch addresses this bug.

This commit was SVN r24626.

The following Trac tickets were found above:
  Ticket 2775 --> https://svn.open-mpi.org/trac/ompi/ticket/2775