Fixes trac:2785: Fix orterun handling of prefix conflcts
Based on the patch attached to the ticket.

r24657
Check for both absolute-path-to-mpirun and -prefix being specified. If the two differ, print out a warning and ignore -prefix. If they are the same, or only one was given, then proceed as directed.

r24658
Grr...get the prefix right - need to strip the bin out of absolute path to mpirun.

This commit was SVN r24693.

The following SVN revision numbers were found above:
  r24657 --> open-mpi/ompi@6af2677fb86393b333205ab591024e4b712480bc
  r24658 --> open-mpi/ompi@0ff0d20e721337c7bad18e75dad20d5d47f8fedf

The following Trac tickets were found above:
  Ticket 2785 --> https://svn.open-mpi.org/trac/ompi/ticket/2785