Fixes trac:2699: Move new MPI-2.2 datatypes and datatype engine improvements to v1.5
r24247
Reshape the datatype engine. The basic types are built down in OPAL. MPI types are
either direct link to these basic predefined types, or a combination of them.
Anyway, the first items in the datatype list belong to OPAL, the second round
are MPI datatypes created by composing basic OPAL datatypes, and the last
batch are mapped datatype (direct correspondance between an OMPI datatype and
an OPAL one such as int -> int32_t).

Modify the op to fit this new scheme.

r24252
Update the tests to match the new datatype engine.

r24279
Correctly initialize the convertor to be used.
Don't forget to initialize the OPAL datatype module.

r24466
Allow the one sided components to correctly retrieve the op to
be applied. Correct the MPI validation process of the
MPI_Accumulate arguments.

Fix another potential problem not yet reported. If we convert the
MPI datatypes direclty into OPAL datatypes, we will restrict their
number to the locally different types. Which might not be identical
on the remote node, if we are in a heterogeneous environment. So,
for MPI One sided only deal with MPI level types, never simplify
them on OPAL types (at least on the args). The unfortunate
outcome is that we need to create the args for all datatypes.

This commit was SVN r24695.

The following SVN revision numbers were found above:
  r24247 --> open-mpi/ompi@5390fd6f33e5df474422d4287119cdf10b3510c3
  r24252 --> open-mpi/ompi@29c7f2fba55b7b666af971f2af48e702c169ef30
  r24279 --> open-mpi/ompi@fc9133cc7f4a811cf05f4bcc6b0b1fbb1d4dcb6c
  r24466 --> open-mpi/ompi@27fecda12cf51891f83b7d91d87fffd393d67e48

The following Trac tickets were found above:
  Ticket 2699 --> https://svn.open-mpi.org/trac/ompi/ticket/2699