Fixes trac:2734: Move asm fix to v1.5
r24449
Fix register clobber list for x86 assembly.  Thanks to Jay Fenlason for the
patch.

r24453
Correct the assembly using xaddl for IA32.
Add atomic functions for add and sub 32 and 64 bits for AMD64.

r24528
Solve a problem when cross-compiling for PPC32 - in this case, OPAL_HAVE_ATOMIC_CMPSET_64 is not set, but the code requires that the ADD_64 and SUB_64 values at least be defined.

This commit was SVN r24931.

The following SVN revision numbers were found above:
  r24449 --> open-mpi/ompi@07996af388379bc2e2a9fe85c65fe4febd2dc3d9
  r24453 --> open-mpi/ompi@f79c87f0c3d260c079a2906aaf574f50413d5517
  r24528 --> open-mpi/ompi@7eede54b39c8ab5716799fcbb02c60b1d13248f0

The following Trac tickets were found above:
  Ticket 2734 --> https://svn.open-mpi.org/trac/ompi/ticket/2734