Closes trac:2831
Oracle would like r24839 brought over to the 1.4 branch. Actually the code that was modified for #2816 is really what we want. I will attach a patch with that code base for 1.4.

We would like to bring this over as soon as reasonable since it stops Oracle from running MTT tests on S11 machines.

This commit was SVN r24953.

The following SVN revision numbers were found above:
  r24839 --> open-mpi/ompi@8c0af7838ac275c72c0dd3f452c2b2ac7ff355b4

The following Trac tickets were found above:
  Ticket 2831 --> https://svn.open-mpi.org/trac/ompi/ticket/2831