Remove the 'ess_node_rank' accessor from here. This caused running under 'tm' to segv at the orteds.
It just looks like this part of the component was not updated during r25331. It was removed from the 'env' and 'slurm' environments in that patch. It looks like 'tm' was updated, but did not get this particular piece.

This commit was SVN r25380.

The following SVN revision numbers were found above:
  r25331 --> open-mpi/ompi@b44f8d4b2848ba54c90dbb617faacb69c43d0167