Fixes trac:2918: Pack error with partial memcopy
---svn-pre-commit-ignore-below---
r25488
Fix an issue with the datatype engine where if the left over from
the previous pack was larger than the allowed space in the pack buffer
bad things happened. Thanks to Yuki MATSUMOTO and Takahiro Kawashima
for the bug report and the patch.

This commit was SVN r25532.

The following SVN revision numbers were found above:
  r25488 --> open-mpi/ompi@3460631fe085a663dcc17ba49688e79a7d3eae6e

The following Trac tickets were found above:
  Ticket 2918 --> https://svn.open-mpi.org/trac/ompi/ticket/2918