Fixes trac:2923: MPI_IN_PLACE fix for v1.4
---svn-pre-commit-ignore-below---
Custom patch on the ticket, based on r25545:

Per http://www.open-mpi.org/community/lists/users/2011/11/17862.php,
to make MPI_IN_PLACE (and other sentinel Fortran constants) work on OS
X, we need to use the following compiler (linker) flag:

    -Wl,-commons,use_dylibs 

So if we're compiling on OS X, test to see if that flag works with the
compiler.  If so, add it to the wrapper FFLAGS and FCFLAGS (note that
per a future update, we'll only have one Fortran compiler anyway).

This commit was SVN r25602.

The following SVN revision numbers were found above:
  r25545 --> open-mpi/ompi@7f9ae11fafc74d1c5f0764eecf13dd4bd3d7d47b

The following Trac tickets were found above:
  Ticket 2923 --> https://svn.open-mpi.org/trac/ompi/ticket/2923