Fixes trac:2944: CMR latest totalview/debuggers fixes from trunk
---svn-pre-commit-ignore-below---
r25713
My apologies for doing this outside of the usual time restrictions, but we need to get this in so we can make progress.

Move the ORTE-level debugger code back into orterun and out of the ORTE library to resolve symbol conflicts.

This commit was SVN r25719.

The following SVN revision numbers were found above:
  r25713 --> open-mpi/ompi@bf103de66c37cb2cc48313d74d351a1e7ba44246

The following Trac tickets were found above:
  Ticket 2944 --> https://svn.open-mpi.org/trac/ompi/ticket/2944