Refs trac:3001.
Should never use AC_MSG_ERROR in a component configure.m4.

This also still needs to be reviewed, so I'm sticking it as part of
number #3001, which Brian will review shortly.  But I'm committing it
now so that we can build 1.5 nightly tarballs.

This commit was SVN r25900.

The following Trac tickets were found above:
  Ticket 3001 --> https://svn.open-mpi.org/trac/ompi/ticket/3001