Wow.  This bug has existed for '''YEARS''' (probably ever since Open MPI started!).
The FLAG argument to fortran attribute copy functions is a LOGICAL,
meaning that it can only return .TRUE. or .FALSE.  The corresponding C
argument is an int, and the MPI spec says that it must return 1 or 0.
However, in Fortran, .TRUE. is not always necessarily == 1.  So we
need to expand the test to see if it's a Fortran callback.  If so,
check for the Fortran .TRUE. value (not 1).  If it's a C callback,
then check for 1.

This commit was SVN r25965.