Fixes trac:2996: Fix default hostfile MCA param
---svn-pre-commit-ignore-below---
r25851
In accordance with prior releases, we are supposed to default to looking at the openmpi-default-hostfile as a default hostfile. Restore that behavior, but ignore the file if it is empty. Allow the user to ignore any MCA param setting pointing to a default hostfile by setting the param to "none" (via cmd line or whatever) - this allows them to override a setting in the system default MCA param file.

This commit was SVN r26217.

The following SVN revision numbers were found above:
  r25851 --> open-mpi/ompi@a5e4dc68030e4c5dbbf6bf0067c7e57afecd886e

The following Trac tickets were found above:
  Ticket 2996 --> https://svn.open-mpi.org/trac/ompi/ticket/2996