Fixes trac:3068: Move fakeroot ptmalloc pre-main hook fix to v1.6
---svn-pre-commit-ignore-below---
r26203
Fixes trac:2812: check for env. markers indicating that we're in a
fakeroot.  If so, exit out of the pre-main hook immediately (without
calling functions such as stat, which will be replaced by fakeroot to
things that are not safe to call in a pre-main environment).

r26204
Using the right env variable name helps!

This commit was SVN r26231.

The following SVN revision numbers were found above:
  r26203 --> open-mpi/ompi@8a2df3311dbaf6551586c353bae745f615f71482
  r26204 --> open-mpi/ompi@028f471a206612e4bea32b21e5d724875aea072f

The following Trac tickets were found above:
  Ticket 2812 --> https://svn.open-mpi.org/trac/ompi/ticket/2812
  Ticket 3068 --> https://svn.open-mpi.org/trac/ompi/ticket/3068