Fixes trac:3005: Move r25886 to v1.6 branch (comm spawn error fix)
---svn-pre-commit-ignore-below---
r25886
Ensure we don't segfault if comm_spawn_multiple fails as it returns a NULL pointer to the new communicator when that happens. Thanks to Yuki MATSUMOTO of Fujitsu for the patch.

cmr:v1.6

This commit was SVN r26296.

The following SVN revision numbers were found above:
  r25886 --> open-mpi/ompi@30ab3b2332cb9087537b8ed5180e405344373968

The following Trac tickets were found above:
  Ticket 3005 --> https://svn.open-mpi.org/trac/ompi/ticket/3005