Oops - if we don't want the Java bindings, then we really shouldn't be building them :-/
Also ensure we don't try to build them if no Java support was found, and error out if the user requests the bindings and we didn't find Java support.

Add a configure flag to skip the Java tests and just force-set the Java support to "disabled"

This commit was SVN r26484.