Fixes trac:3186: CMR: change default alltoall algorithm in 1.6
---svn-pre-commit-ignore-below---

r26246 [[BR]]
As suggested on the mailing list a while back, switch the default alltoallv
algorithm to pairwise exchange instead of the default one. This might
improve the scheduling and relax the pressure on the network.

This commit was SVN r26890.

The following SVN revision numbers were found above:
  r26246 --> open-mpi/ompi@654c75ff24cda4eb1419b9a6d700387237562f63

The following Trac tickets were found above:
  Ticket 3186 --> https://svn.open-mpi.org/trac/ompi/ticket/3186