Fixes trac:3203: Fix rsh termination race condition
---svn-pre-commit-ignore-below---

r26935 [[BR]]
It's actually ''not'' an error to exit if all routes and children are
gone.  So exit with 0, not ORTE_ERROR_DEFAULT_EXIT_CODE (which is 1).

This fixes a race condition in the rsh launcher upon termination,
where ORTE would sometimes think that a daemon failed to launch.

This commit was SVN r26937.

The following SVN revision numbers were found above:
  r26935 --> open-mpi/ompi@62c2ff7ee7252023dae98b69130c30f6131f9235

The following Trac tickets were found above:
  Ticket 3203 --> https://svn.open-mpi.org/trac/ompi/ticket/3203