Fixes trac:3224: Add a couple of useful states for 3rd party tools
---svn-pre-commit-ignore-below---
r26716
Add a state that allows mpirun or other tools to be notified of a job completion prior to terminating so that alternative actions can be performed.

r26902
Add an intermediate state when the VM is ready so that third party tools can take action prior to mapping/launching apps

r26947
If a callback is set in the state machine, and the state doesn't yet exist, create it

This commit was SVN r27008.

The following SVN revision numbers were found above:
  r26716 --> open-mpi/ompi@b83fc41d548504cde6ca0916875741eeffc54b66
  r26902 --> open-mpi/ompi@94d11e04fd5d825c00e3ba4b4e873b3405ee566a
  r26947 --> open-mpi/ompi@0d878937fefce8c68e6ae01fdc0d36984bbee2c1

The following Trac tickets were found above:
  Ticket 3224 --> https://svn.open-mpi.org/trac/ompi/ticket/3224