Fixes trac:3208: Update ctrl-c behavior in v1.7
---svn-pre-commit-ignore-below---
r26943
Ensure we can forcibly exit even when we are stuck inside of an event by replacing the libevent signal handler with a POSIX one that (a) attempts to trip a libevent termination event and (b) if anothe ctrl-c hits within 5 seconds, just calls exit.


r26945
A little cleaner on the message about repeated ctrl-c, and re-enable the event so we can abort if we see multiple ctrl-c's that don't meet the time requirement

r27005
Just a piece of this changeset - delete term_handler event upon normal termination.

This commit was SVN r27011.

The following SVN revision numbers were found above:
  r26943 --> open-mpi/ompi@e6c72bfd5320c6bc08c3e4a44931b95ad2cb8412
  r26945 --> open-mpi/ompi@dc22ea5cde6154e706637954d7531a8ed25bd69c
  r27005 --> open-mpi/ompi@cb48fd52d4f355d3a2a1e9cca4853a198d059871

The following Trac tickets were found above:
  Ticket 3208 --> https://svn.open-mpi.org/trac/ompi/ticket/3208