Fixes trac:3234: Fix nidmap decode
---svn-pre-commit-ignore-below---
r27035
We can't just lookup the node in the node pool by daemon vpid as the daemons aren't stored that way - this was done because when holes exist in daemon vpids, we can generate huge orte_node_pool arrays even when only a few daemons actually exist. So we have to search for the vpid in the array

r27038
This will break systems like orcm, but we aren't trying to support those any more - so put the nodes back in their daemon-indexed position. Will continue working to reduce search requirements in other parts of the code

This commit was SVN r27058.

The following SVN revision numbers were found above:
  r27035 --> open-mpi/ompi@3cb8d55c8be9ce15ebebd0f3b991315780a3978c
  r27038 --> open-mpi/ompi@229e3f9b2aaba904b97cbd510d6c1267acdd1971

The following Trac tickets were found above:
  Ticket 3234 --> https://svn.open-mpi.org/trac/ompi/ticket/3234