function mca_btl_openib_endpoint_post_send can return 3 statuses: - OMPI_SUCCESS - OMPI_ERROR - OMPI_ERR_RESOURCE_BUSY
If an "OMPI_ERR_OUT_OF_RESOURCE" occurs, the request is added to the pending list, and will be handled later. An error message 
should not be printed to the user in this case. This is not an error, but rather a notification of a possible valid condition.
Only in the case of "OMPI_ERROR" should it be printed to the user.

This commit was SVN r27065.