Refs trac:3248
Absoft has determined exactly what the problem is (private members in
derived data typed when compiled with -g), but does not yet have a
timeline for fixing it.

Add a temporary override for Absoft until they are able to fix their
compiler.  This switch will at least allow us to MTT test the rest of
the mpi_f08 functionality with Absoft.

This commit was SVN r27184.

The following Trac tickets were found above:
  Ticket 3248 --> https://svn.open-mpi.org/trac/ompi/ticket/3248