Refs trac:3293: Cleanup hostfile behavior in managed environments
---svn-pre-commit-ignore-below---
Add the following to #3293 -- it's documentation only, so it doesn't
affect any code paths.

r27264 [[BR]]
Add a note in the default hostfile that it is not used in managed
environments. 

This commit was SVN r27265.

The following SVN revision numbers were found above:
  r27264 --> open-mpi/ompi@8585920e49276024e236c675bfce6275e1ecf255

The following Trac tickets were found above:
  Ticket 3293 --> https://svn.open-mpi.org/trac/ompi/ticket/3293