Fixes trac:3287: Update hostfile behavior
---svn-pre-commit-ignore-below---

r27223 [[BR]]
Provide a more user-expected way of handling hostfile and dash-host allocations in unmanaged environments. First look for an RM-managed allocation. If nothing is found, or no active module is alive, proceed to look for hostfile and dash-host assignments. These are provided on a per-app basis, so we have to cycle across the apps using the following algorithm:

1. if a hostfile is given, then add the nodes found in that hostfile to our list - i.e., the resulting allocation contains the UNION of all nodes specified in hostfiles from across all apps.

2. any app that has no hostfile but has a dash-host, will have those nodes added to the list

3. any app that fails to have a hostfile or a dash-host will be given the default hostfile, if we have it

Each app will subsequently be filtered using their hostfile and/or dash-host data to ensure that the app only has access to the hosts it specified

Note that any relative node syntax found in the hostfiles or dash-host data will generate an error in this scenario, so only non-relative syntax can be present


r27230 [[BR]]
Complete the revamp of hostfile support in non-managed environments. Working at the app level, ensure that we utilize only those nodes specified for that app, but fall back to the default hostfile (if available) for those with no specification, further falling back to the local host if the default hostfile is not present or is empty.


r27231 [[BR]]
Move the debug output to the right place

This commit was SVN r27294.

The following SVN revision numbers were found above:
  r27223 --> open-mpi/ompi@42d58f17bf5ca590d31736966441351348106758
  r27230 --> open-mpi/ompi@11de735e8a9044e78680ec727c16ea5234f4e96a
  r27231 --> open-mpi/ompi@86a16edd5a72c62778009dae452d3c80f6cb4ee0

The following Trac tickets were found above:
  Ticket 3287 --> https://svn.open-mpi.org/trac/ompi/ticket/3287