Fixes trac:3313: Fix multi-node comm_spawn in v1.7
---svn-pre-commit-ignore-below---

r27318 [[BR]]
Cleanup comm_spawn for the multi-node case where at least one new process isn't spawned on every node. Avoid the complexities of trying to execute a daemon collective across the dynamic spawn as it becomes too hard to ensure that all daemons participate or are accounted for - instead, use a less scalable but workable solution of sending the data directly between the participating procs. Ensure that singletons get their collectives properly defined at startup so the spawned "HNP" is ready for them.

As a secondary cleanup, the HNP doesn't need to update its nidmap during an xcast as it already has an up-to-date picture of the situation. So just dump that data and move along.

This commit was SVN r27327.

The following SVN revision numbers were found above:
  r27318 --> open-mpi/ompi@c82cfecc1ca3035d8af680ac39b667be6aca8800

The following Trac tickets were found above:
  Ticket 3313 --> https://svn.open-mpi.org/trac/ompi/ticket/3313