Fixes trac:3322: Move r27342 to v1.7 branch
---svn-pre-commit-ignore-below---

r27342 [[BR]]
Fix the bind_downward code - it was incorrectly looking across the entire node instead of only looking below the locale to which the proc had been assigned. In other words, if the proc was mapped to a core, then the only hwthreads that should be considered for binding are those directly below that core. The binding algo was incorrectly looking at ALL hwthreads in that scenario, causing the proc to be bound to an HT outside of the mapped location.

This now results in the procs being bound within their assigned location. It also causes us to use only the 0th HT on a core unless --use-hwthread-cpus has been specified (in which case, we use all the HTs in a core). Bind to core binds you to all HTs regardless - the --use-hwthread-cpus only impacts the oversubscribed determination and when binding to HT.

cmr:v1.7


r27354 [[BR]]
Update test

r27357 [[BR]]
Refs trac:3322

Cannot start the data clearing at the root object level as the root object has a different struct attached to userdata.

This commit was SVN r27408.

The following SVN revision numbers were found above:
  r27342 --> open-mpi/ompi@a3060cdd15347ce22a78863814bf6bd940e079c8
  r27354 --> open-mpi/ompi@90d7b5fdca2d05352fed2a0ff7f37e8448add4d8
  r27357 --> open-mpi/ompi@662bc05aa61e0e734434a23f04ef1124cb8732b3

The following Trac tickets were found above:
  Ticket 3322 --> https://svn.open-mpi.org/trac/ompi/ticket/3322