Fixes trac:3328: Move TSD change in memory/linux to 1.7
---svn-pre-commit-ignore-below---

r27371 [[BR]]
Remove the TSD_HACKS macro.  The TSD hack is only for non-glibc libraries
and we only build the linux memory component on glibc, so this shouldn't
be needed.

This commit was SVN r27409.

The following SVN revision numbers were found above:
  r27371 --> open-mpi/ompi@cb6831830ac50d504a8bdf2dc28658b790dcc879

The following Trac tickets were found above:
  Ticket 3328 --> https://svn.open-mpi.org/trac/ompi/ticket/3328