Fixes trac:3345: Please move r27400 to 1.7
GK note: had to resolve a conflict in the #include area - edited to agree with the trunk
---svn-pre-commit-ignore-below---

r27400 [[BR]]
Fixing a coupe of issues in the iboffload code. It should fix the XRC compilation issue.

This commit was SVN r27414.

The following SVN revision numbers were found above:
  r27400 --> open-mpi/ompi@e9df33c10b9221dbfd3c1a95c769e2f1408844b6

The following Trac tickets were found above:
  Ticket 3345 --> https://svn.open-mpi.org/trac/ompi/ticket/3345