Revert r27510, r27509, and r27508.
Not sure what happened here, but the resulting trunk wouldn't even configure. After spending time fixing that problem, I found it wouldn't compile due to multiple syntax errors that had been introduced in both the OPAL and OMPI layer. This raised questions as to the completeness of the work.

Given that the author is departing, I pinged Jeff about it and we agreed to revert this for now. Hopefully, it can either be fixed by the author prior to actual departure, or someone else can pick it up (now that it is in the history) and fix it.

This commit was SVN r27511.

The following SVN revision numbers were found above:
  r27508 --> open-mpi/ompi@12c3c743dea33b141fe7ca642d015ecc2f46c02a
  r27509 --> open-mpi/ompi@79e4a8ca380da249aa8e61c1d11dd294f02a8b62
  r27510 --> open-mpi/ompi@1ad5ff625a7f665a53d3f1cd73af63beee29b59b