Fixes trac:3367: Move r27522 to v1.7 branch
---svn-pre-commit-ignore-below---

r27522 [[BR]]
Update to r27292: only use "valuelen-1" when the valuelen is
MPI_MAX_INFO_VAL (so that we don't go beyond the end of the string).
But otherwise, abibe by what says in MPI-2.2 / MPI-3.0 in the
description of MPI_INFO_GET (where "it" refers to the valuelen
argument):

    If it is less than the actual size of the value, the value is
    truncated. In C, valuelen should be one less than the amount of
    allocated space to allow for the null terminator.

cmr:v1.7

This commit was SVN r27525.

The following SVN revision numbers were found above:
  r27292 --> open-mpi/ompi@fb4af5e29c2f03746dbdefc5e773f15437d8ca21
  r27522 --> open-mpi/ompi@a980ee5ca126f8df0376f37ddfc57dedb8e77f6a

The following Trac tickets were found above:
  Ticket 3367 --> https://svn.open-mpi.org/trac/ompi/ticket/3367