Revert "Revert r27451 and r27456 - the cmd line parser is incorrectly marking the application as an MCA parameter".
It appears the problem was not with the command line parser but the rsh plm. I don't know why this problem was not occuring before the command line parser changes but it appears to be resolved now.

This commit was SVN r27527.

The following SVN revision numbers were found above:
  r27451 --> open-mpi/ompi@d59034e6efff3aa4904ccfdea891d1178b4d2fdf
  r27456 --> open-mpi/ompi@ecdbf3493781081bc07aac86285a96b867790d36