Fixes trac:3389: Move r27561 to v1.7 branch (retain full FQDN hostnames)
---svn-pre-commit-ignore-below---

r27561 [[BR]]
If the nodename is an IP address, we need to retain the full name (even if keep_fqdn is false) so that the ssh tree spawn can proceed.

cmr:v1.7


r27583 [[BR]]
Add a function to test if a string is in the form of an IP address - doesn't test for validity of the address

r27586 [[BR]]
Strip the domain info from the hostname if that option is specified, protecting IP address-based names

r27618 [[BR]]
Add a function to return the aliases (based on IP addrs) for the current node

r27619 [[BR]]
Add an MCA param to retain all aliases based on IP addrs for node names so that procs can look them up by interface, if desired. If the param is set, pass aliases around to all daemons and procs for local use

r27620 [[BR]]
Per Jeff's suggestion, set the default answer when asking for IP aliases in case we don't find any

This commit was SVN r27711.

The following SVN revision numbers were found above:
  r27561 --> open-mpi/ompi@27b41a7db4b7cafc62c52c1437beed4b6fb972fc
  r27583 --> open-mpi/ompi@f9f07e95355d714780c5fbca375556872c309191
  r27586 --> open-mpi/ompi@81d0b068429c9cbd004de59cc051ac863533c000
  r27618 --> open-mpi/ompi@a52071a17d308e71c0afbd530696f9de3796d2f1
  r27619 --> open-mpi/ompi@e11f32038a7c34deda3493d451ec65e9ecddef91
  r27620 --> open-mpi/ompi@fdf7633cff9db8b037a5caa00f5dd668ade06f0e

The following Trac tickets were found above:
  Ticket 3389 --> https://svn.open-mpi.org/trac/ompi/ticket/3389