On OpenBSD (and probably others), expr operates on a 32 bit signed integer, which is the same size as a Fortran or C integer.  This resulted in configure coming up with Fortran's MAX_INT as -2^31, which obviously isn't a positive number.  Since we found the MAX_INT using the same broken loop in a couple places and doing it right is complicated, added a new macro that is much more careful about sign roll-over.
During the Fortran rework between v1.6 and v1.7, the variable which
indicates whether or not Fortran is being compiled changed, so on platforms
without Fortran compilers, we were trying to determine the max value for
Fortran INTEGERS where we previously didn't.  I believe this is why
bug #3374 appeared as a regression.

Finally, since the OMPI code doesn't cope with OMPI_FORTRAN_HANDLE_MAX
being negative (which was the root cause of the segfault in $3374),
add a check at the end of the OMPI_FORTRAN_GET_HANDLE_MAX macro to
ensure that OMPI_FORTRAN_HANDLE_MAX is always non-negative.

This commit was SVN r27714.