Fixes trac:3477: Fix debugger attach
---svn-pre-commit-ignore-below---

r27863 [[BR]]
In order to properly connect to debuggers like STAT, we need to get the hostname in its unstripped version for the MPIR_proctab. Unfortunately, we need a stripped version for Cray's alps launcher. So when we are stripping the hostname prefix, retain alias hostnames and add the ability to specify an alias to use in the proctab.

This commit was SVN r27907.

The following SVN revision numbers were found above:
  r27863 --> open-mpi/ompi@c96cc2d5a0b67d17b13ffae824c438fe7009ae81

The following Trac tickets were found above:
  Ticket 3477 --> https://svn.open-mpi.org/trac/ompi/ticket/3477