Fixes trac:3501: Move minor configure options fix to v1.7
---svn-pre-commit-ignore-below---

Custom patch on ticket originally based on r27998:

r27998 [[BR]]
Can't just use "0" in an AM_CONDITIONAL -- you have to use a valid
executable that can be run and returns a false exit status (which,
amusingly, in this case is 1, not 0).

This commit was SVN r28001.

The following SVN revision numbers were found above:
  r27998 --> open-mpi/ompi@da5d093fdc2ce97a85f0489307fd513a1beb019e

The following Trac tickets were found above:
  Ticket 3501 --> https://svn.open-mpi.org/trac/ompi/ticket/3501