Fixes trac:3232: 2D string array handling broken in Fortran MPI_COMM_SPAWN_MULTIPLE
---svn-pre-commit-ignore-below---

r28164 [[BR]]
Ugh.  Work around an issue with memory hooks and the change from one big
library to multiple libraries that are implicitly sucked into the executable
as a dependency of libmpi.  The initialize hook isn't visible to libc on some
linux distributions when it's in libopal and libopal isn't explicity linked
into the executable.  The fix is to have a duplicate initialize hook in
libmpi as well as libopal.  *sigh*.

This commit was SVN r28169.

The following SVN revision numbers were found above:
  r28164 --> open-mpi/ompi@fc2b3b8d4615849b251ed6a56fff37ca84ab0bef

The following Trac tickets were found above:
  Ticket 3232 --> https://svn.open-mpi.org/trac/ompi/ticket/3232