Fix an issue identified by Thomas Jahns and his colleague when the data representation is not correctly optimized (it is off by the extend).
During the data representation process, if the opportunity to merge several
items appear, we replace them with the new merged element. However, if one
of the components of this merged element was comming from a "loop representation"
then the new first element of this loop must have a displacement moved by the
extent of the loop.