Fixes trac:3559: Move 28319 to v1.7 (datatype fix)
---svn-pre-commit-ignore-below---

r28319 [[BR]]
Fix an issue identified by Thomas Jahns and his colleague when the data
representation is not correctly optimized (it is off by the extend).

During the data representation process, if the opportunity to merge several
items appear, we replace them with the new merged element. However, if one
of the components of this merged element was comming from a "loop representation"
then the new first element of this loop must have a displacement moved by the
extent of the loop.

This commit was SVN r28386.

The following SVN revision numbers were found above:
  r28319 --> open-mpi/ompi@43e4d3654eb3ff2f18f410330022a91afc6e81b7

The following Trac tickets were found above:
  Ticket 3559 --> https://svn.open-mpi.org/trac/ompi/ticket/3559