Fixes trac:3575: OB1 convertor reset fix
---svn-pre-commit-ignore-below---

r28292 [[BR]]
pml/ob1: do not reset the convertor if one was not created (size = 0).

This macro is only used on the failure path so the additional if statement
should not have any affect on performance.

This commit was SVN r28393.

The following SVN revision numbers were found above:
  r28292 --> open-mpi/ompi@4e95d691a7e2d4354586a64578f128cf6a947c13

The following Trac tickets were found above:
  Ticket 3575 --> https://svn.open-mpi.org/trac/ompi/ticket/3575